droplet/requirements.txt
Django==1.9
django-admin-bootstrapped==2.5.7
django-authtools==1.5.0
django-bootstrap-form==3.2.1
django-braces==1.9.0
django-contrib-comments==1.7.3
django-cors-headers==1.3.1
django-crispy-forms==1.6.1
django-crudbuilder==0.2.5
django-debug-toolbar==1.6
django-environ==0.4.0
django-extensions==1.7.5
django-filter==1.0.1
django-helpdesk==0.1.18
django-markdown-deux==1.0.5
django-tables2==1.3.0
django-threadedcomments==1.0.1
django-tinymce==2.4.0
djangorestframework==3.5.3
djangorestframework-jwt==1.9.0
easy-thumbnails==2.3
email-reply-parser==0.5.9
markdown2==2.3.2
Pillow==3.4.1
py==1.4.31
PyJWT==1.4.2
PyMySQL==0.7.9
simplejson==3.10.0
six==1.10.0
sqlparse==0.2.1
Werkzeug==0.11.11
